My multicast protocol should be killed with fire but it works nonetheless!

+++ Server/Server.cpp
@@ -43,19 +43,13 @@ ServerState Server::getState()
 
 void Server::setState(ServerState serverState)
 {
-	m_state.store(serverState);
+	m_state.store(static_cast<const byte>(serverState));
 #ifdef SERVER_LOGGING
 	switch (serverState)
 	{
 	case ServerState::IDLE:
 		printf("Idling...\n");
 		break;
-	//case ServerState::CONNECT:
-	//	printf("Connecting.\n");
-	//	break;
-	//case ServerState::DISCONNECT:
-	//	printf("Disconnecting.\n");
-	//	break;
 	case ServerState::ROUTE:
 		printf("Routing.\n");
 		break;
@@ -71,10 +65,6 @@ void Server::run()
 		switch (getState()) {
 		case ServerState::IDLE:
 			break;
-		//case ServerState::CONNECT:
-		//	break;
-		//case ServerState::DISCONNECT:
-		//	break;
 		case ServerState::ROUTE: {
 			std::future<void> syncRecv = std::async(&Server::recvAll, this);
 			std::future<void> syncSend = std::async(&Server::sendAll, this);
